Add create sc storage costs and add some tests.
